requirements.txt
geojson==3.1.0
geopandas==0.9.0
geopy==2.4.1
h5netcdf==1.1.0
lightning==1.9.5
lightning-cloud==0.5.70
lightning-utilities==0.8.0
pytorch-lightning==2.0.2
seaborn==0.12.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchgeo==0.4.1
torchmetrics==0.11.4
torchvision==0.15.2+cu118
tqdm==4.65.0
xarray==2023.1.0